+++ src/mapleall/maple_be/src/be/switch_lowerer.cpp
@@ -111,7 +111,7 @@ RangeGotoNode *SwitchLowerer::BuildRangeGotoNode(int32 startIdx, int32 endIdx) {
      */
     while ((stmt->GetCasePair(i).first != (lastCaseTag + 1)) && (stmt->GetCasePair(i).first != lastCaseTag)) {
       /* fill in a gap in the case tags */
-      curTag = (++lastCaseTag) - node->GetTagOffset();
+      curTag = static_cast<uint32>(static_cast<int32>(++lastCaseTag) - node->GetTagOffset());
       if (stmt->GetDefaultLabel() != 0) {
         node->AddRangeGoto(curTag, stmt->GetDefaultLabel());
       }
